third_party/WebKit/LayoutTests/external/wpt/css/css-scroll-snap/snap-at-gesture-scroll-end-manual.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1" />
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<style>
body {
  margin: 0px;
  overflow: scroll;
  scroll-snap-type: both mandatory;
}
#content {
  width: 2000px;
  height: 2000px;
  padding: 0px;
  margin: 0px;
}
#target {
  position: relative;
  left: 400px;
  top: 400px;
  width: 400px;
  height: 400px;
  border: red solid 5px;
  scroll-snap-align: start;
}
#i1 {
  color: red;
  font-weight: bold;
}
</style>

<div id="content">
  <div id="target">
    <h1>CSSScrollSnap</h1>
    <h4>Tests that the window can snap at GestureScrollEnd</h4>
    <ol>
      <li id="i1" style="color: red">
        Scroll the page vertically and horizontally in one gesture.
        Release after the background has turned yellow.</li>
      <li id="i2"> Press the button "Done"</li>
    </ol>
    <input type="button" id="btn" value="DONE" style="width: 100px; height: 50px;"/>
  </div>
</div>

<script>
  var body = document.body;
  var button = document.getElementById("btn");
  var instruction1 = document.getElementById("i1");
  var instruction2 = document.getElementById("i2");
  var scrolled_x = false;
  var scrolled_y = false;
  var start_x = window.scrollX;
  var start_y = window.scrollY;
  var test = async_test("Window should snap at GestureScrollEnd");

window.onscroll = function() {
  if (scrolled_x && scrolled_y) {
    body.style.backgroundColor = "yellow";
    instruction1.style.color = "black";
    instruction1.style.fontWeight = "normal";
    instruction2.style.color = "red";
    instruction2.style.fontWeight = "bold";
    return;
  }

  if (!scrolled_x && window.scrollX != start_x)
    scrolled_x = true;
  if (!scrolled_y && window.scrollY != start_y)
    scrolled_y = true;
}

button.onclick = function() {
  if (!scrolled_x || !scrolled_y)
    return;

  test.step(function() {
    assert_equals(window.scrollX, 400);
    assert_equals(window.scrollY, 400);
    var content = document.getElementById("content");
    body.removeChild(content);
    test.done();
  });
}

</script>
